+++ chat21/src/main/java/org/chat21/android/core/contacts/synchronizers/ContactsSynchronizer.java
@@ -103,7 +103,7 @@ public class ContactsSynchronizer {
                     }.execute(dataSnapshot);
                 }
 
-                //for return recepit
+                //for return receipt
                 @Override
                 public void onChildChanged(DataSnapshot dataSnapshot, String prevChildKey) {